fix locale ref error
